Add averaging count variables to default output streams
This commit is necessary for restart history files, which need to also
load the averaging count for each file where filled variables may be
present.  This commit also adds the name of the averaging count to
a variables metadata so in post-processing the user can check the
values against the averaging count.