Add helper field functions to nudging process similar to surface export
This commit copies the syntax used in surface exporting to facilitate
a set of helper fields to be used in the nudging process.

A task for future development is to consilidate the internal fields and
these helper fields into a single object in an atmosphere process, which
will then replace this syntax.  But that task is outside the scope of this
work so we will use the helper fields for now.