Accumulate data land-ice mass fluxes
Before this merge, we were only accumulating this flux in
prognostic flux mode.  With this merge, we accumulate the mass
flux in data mode as well, as this term is needed in the total
mass budget of the model.

To facilitate this and reduce issues related to initialization
vs. forward time stepping, this accumulation step has been broken
into its own subroutine, ocn_surface_land_ice_fluxes_accumulate_fluxes.
This subroutine is called when time-stepping the model but not
at initialization (since we do not want to accumulate any fluxes
at that time).