Fixes an inconsistency between standard IO and vertical remap files.
This commit fixes an inconsistency between the typical EAMxx output
which uses the dimension name `lev` for levels and the vertical
remapping files which uses `nlevs`.  This isn't a major bug, but the
inconsistency can cause some confusion and more importantly, if someone
uses EAMxx output for nudging they would have a pre-processing step
of adding a dimension `levs` to the file containing `p_levs` just
to work.