Add a nudging standalone test to the test suite.
This commit adds the test shoc_p3_nudging to the set of "stand-alone"
unit tests that run through the AD.  This test is similar to the restart
test in the same suite in that it runs two simulation.  The first generates
nuding data while the second runs w/ nudging turned on.

While making this test a bug was uncovered, that isn't addressed in this test
but will be fixed in a subsequent commit.  The bug is in cases where the number
of levels isn't a multiple of the packsize, the vertical interpolation isn't a
fan.