EAMxx: use gid_type instead of gid_t
The latter is a type in sys/types.h, and it is an unsigned 32bit int.
A few places were using gid_t, likely assuming it was coming from something like
  using gid_t = AbstractGrid::gid_type
That wasn't the case. It caused some warnings, and was a potential
source of bugs should gid_t have a different size on a new machine.