clean-clear-sky and clean-sky radiation calls
Adds two additional radiation diagnostic calls following the logic of clear-sky calls. These are clean-clear-sky (neither clouds nor aerosols) and clean-sky (no aerosols). The latter in enabled by instantiating a second optics and never endowing it with aerosols, while the former is called before the other optics are endowed with aerosols and clouds.