Fix issue with avg cnt tracking, horiz remapping  and 2D sliced output
This commit fixes an issue where requesting 2D sliced output, like
XXX_at_NNNhPa would lead to an FPE when horizontal remapping was used
and output was not INSTANT.

The issue was related to the mechanism for tracking the number of
unmasked values that contributed to the average.  IO creates an internal
field call avg_cnt_tracking_BLAH that is meant to tally how many snaphots
on the layout of the output have contributed to the running average.

The bug is in the layout assigned to this internal field.  It was based on
the simulation grid rather than the output grid.  When horizontal remapping
was requested this meant that the avg cnt tracking was meant to be on a
different grid then it was assigned and thus an FPE would trigger.