wip - Allow polygonal landunits to be initiated from surface file
Updates initialization routines to allow for polygonal landunits
to be created from namelist option and from polygon type
percentages specified on input file. In particular,
it allows for four landunits to be created with istsoil ltype
on each topounit so that the ice wedge polygons can have their
own soil columns as well as be vegetated. The fourth landunit
is for non-polygonal natural vegetated ground on the same topounit.